requirements.txt
absl-py==1.4.0
accelerate==0.19.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
art==6.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
Authlib==1.2.1
Babel==2.12.1
bangla==0.0.2
bardapi==0.1.23a0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
boltons==23.0.0
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid==0.1.35
click==7.1.2
clip-anytorch==2.5.2
cloudinary==1.40.0
colorama==0.4.6
comm==0.2.2
contourpy==1.0.7
coqpit==0.0.17
cryptography==41.0.4
cssselect==1.2.0
cssutils==2.7.1
cvzone==1.5.6
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.8
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
discord.py==2.3.2
distlib==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
einops==0.6.1
email-to==0.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.20.0
ffmpy==0.3.1
filelock==3.12.0
firebase==4.0.1
firebase-admin==6.2.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
flatbuffers==24.3.25
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
g2pkk==0.1.2
gast==0.6.0
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.4.0
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gradio==4.7.1
gradio_client==0.7.0
greenlet==1.0.0
grpcio==1.56.2
grpcio-status==1.56.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
gunicorn==20.0.4
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.14.1
hyperframe==6.0.1
idna==3.4
imageio==2.28.1
img2pdf==0.5.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflect==5.6.0
ipykernel==6.29.5
ipython==8.18.1
isoduration==20.11.0
itsdangerous==1.1.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.2.0
json5==0.9.25
jsonlines==1.2.0
jsonmerge==1.9.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
k-diffusion==0.0.15
keras==3.4.1
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
kornia==0.6.12
lazy_loader==0.2
libclang==18.1.1
librosa==0.10.0.post2
llvmlite==0.39.1
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mecab-python3==1.0.5
mediapipe==0.8.10
mistune==3.0.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mysql-connector-python==8.2.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.8.1
notebook_shim==0.2.4
num2words==0.5.12
numba==0.56.4
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
orjson==3.9.10
outcome==1.2.0
overrides==7.7.0
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pikepdf==8.11.2
pillow==10.2.0
pipenv==2023.5.19
platformdirs==3.5.1
pooch==1.6.0
premailer==3.10.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.22.3
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.16.1
PyJWT==2.8.0
PyMsgBox==1.0.9
pynndescent==0.5.10
pyparsing==3.1.0
pyperclip==1.8.2
pypinyin==0.49.0
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.19.3
pysbd==0.3.4
PyScreeze==0.1.30
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
python-telegram-bot==20.3
pytweening==1.0.7
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0
pyzmq==26.0.3
referencing==0.35.1
regex==2023.5.5
requests==2.31.0
resize-right==0.0.2
responses==0.24.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.19.0
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
secure-smtplib==0.1.1
selenium==4.12.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==1.23.1
setproctitle==1.3.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.5
SQLAlchemy==2.0.21
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.3.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.15.2
tornado==6.4.1
tqdm==4.65.0
trainer==0.0.20
traitlets==5.14.3
trampoline==0.1.2
transformers==4.29.2
trio==0.22.2
trio-websocket==0.10.4
TTS==0.14.0
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
umap-learn==0.5.1
unidic-lite==1.0.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.24.0.post1
vidstream==0.0.14
virtualenv==20.23.0
virtualenv-clone==0.5.7
wandb==0.15.3
wcwidth==0.2.6
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==1.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0